+++ duva/tests/client_ops/mod.rs
@@ -1,0 +1,8 @@
+mod test_config_get_dir;
+mod test_del;
+mod test_exists;
+mod test_idempotent_client;
+mod test_keys;
+mod test_replication_info;
+mod test_set_get;
+mod test_snapshot_persists_and_recovers_state;
