+++ duva/tests/client_ops/test_config_get_dir.rs
@@ -1,12 +1,11 @@
 /// if the value of dir is /tmp, then the expected response to CONFIG GET dir is:
 /// *2\r\n$3\r\ndir\r\n$4\r\n/tmp\r\n
-mod common;
 use std::time::Duration;
 
-use crate::common::array;
-use common::{ServerEnv, spawn_server_process};
 use duva::clients::ClientStreamHandler;
 use tokio::time::sleep;
+
+use crate::common::{ServerEnv, array, spawn_server_process};
 
 #[tokio::test]
 async fn test_config_get_dir() {
