+++ duva/tests/client_ops/test_del.rs
@@ -2,13 +2,11 @@
 /// Firstly, we set a key with a value and an expiry of 300ms
 /// Then we get the key and check if the value is returned
 /// After 300ms, we get the key again and check if the value is not returned (-1)
-mod common;
-use common::{ServerEnv, array, spawn_server_process};
-
 use duva::{clients::ClientStreamHandler, domains::query_parsers::query_io::QueryIO};
 
+use crate::common::{ServerEnv, array, spawn_server_process};
 #[tokio::test]
-async fn test_exists() {
+async fn test_del() {
     // GIVEN
     let env = ServerEnv::default();
     let process = spawn_server_process(&env);
@@ -23,14 +21,15 @@ async fn test_exists() {
         QueryIO::SimpleString("OK RINDEX 2".into()).serialize()
     );
 
-    // WHEN & THEN
+    // WHEN - set key with expiry
     assert_eq!(
-        h.send_and_get(&array(vec!["exists", "a", "c", "d"])).await,
+        h.send_and_get(&array(vec!["del", "a", "c", "d"])).await,
         QueryIO::SimpleString("2".into()).serialize() // 2 means 2 keys deleted
     );
 
-    assert_eq!(
-        h.send_and_get(&array(vec!["exists", "x"])).await,
-        QueryIO::SimpleString("0".into()).serialize() // 2 means 2 keys deleted
-    );
+    assert_eq!(h.send_and_get(&array(vec!["get", "a"])).await, QueryIO::Null.serialize());
+
+    // THEN
+    let res = h.send_and_get(&array(vec!["GET", "somanyrand"])).await;
+    assert_eq!(res, QueryIO::Null.serialize());
 }
