+++ duva/tests/client_ops/test_snapshot_persists_and_recovers_state.rs
@@ -1,12 +1,5 @@
-/// Cache config should be injected to the handler!
-/// This is to enable client to configure things dynamically.
+use crate::common::{ServerEnv, array, spawn_server_process};
 
-/// if the value of dir is /tmp, then the expected response to CONFIG GET dir is:
-/// *2\r\n$3\r\ndir\r\n$4\r\n/tmp\r\n
-mod common;
-use crate::common::array;
-use common::ServerEnv;
-use common::spawn_server_process;
 use duva::clients::ClientStreamHandler;
 use duva::domains::query_parsers::query_io::QueryIO;
 use std::time::SystemTime;
@@ -14,7 +7,7 @@ use std::time::UNIX_EPOCH;
 
 // TODO response cannot be deterministic!
 #[tokio::test]
-async fn test_save_read_snapshot() {
+async fn test_snapshot_persists_and_recovers_state() {
     // GIVEN
     let env = ServerEnv::default().with_file_name(create_unique_file_name("test_save_dump"));
     let leader_process = spawn_server_process(&env);
@@ -38,39 +31,31 @@ async fn test_save_read_snapshot() {
     // check replication info
     let res = h.send_and_get(&array(vec!["INFO", "replication"])).await;
     let info: Vec<&str> = res.split("\r\n").collect();
-    let prev_leader_repl_id = info[3].split(":").collect::<Vec<&str>>()[1];
-    let prev_leader_repl_offset = info[4].split(":").collect::<Vec<&str>>()[1];
 
-    // THEN
+    // WHEN
     assert_eq!(h.send_and_get(&array(vec!["SAVE"])).await, QueryIO::Null.serialize());
 
     // wait for the file to be created
     tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
-
-    // THEN
     // kill leader process
     drop(leader_process);
 
     // run server with the same file name
-    let leader_process = spawn_server_process(&env);
+    let new_process = spawn_server_process(&env);
 
     // wait for the server to start
     tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
 
-    let mut h = ClientStreamHandler::new(leader_process.bind_addr()).await;
+    let mut client = ClientStreamHandler::new(new_process.bind_addr()).await;
 
-    // keys
-    assert_eq!(h.send_and_get(&array(vec!["KEYS", "*"])).await, array(vec!["foo2", "foo"]));
+    assert_eq!(client.send_and_get(&array(vec!["KEYS", "*"])).await, array(vec!["foo2", "foo"]));
 
     // replication info
-    let res = h.send_and_get(&array(vec!["INFO", "replication"])).await;
-    let info: Vec<&str> = res.split("\r\n").collect();
-    let leader_repl_id = info[3].split(":").collect::<Vec<&str>>()[1];
-    let leader_repl_offset = info[4].split(":").collect::<Vec<&str>>()[1];
+    let res = client.send_and_get(&array(vec!["INFO", "replication"])).await;
+    let info2: Vec<&str> = res.split("\r\n").collect();
 
     // THEN
-    assert_eq!(leader_repl_id, prev_leader_repl_id);
-    assert_eq!(leader_repl_offset, prev_leader_repl_offset);
+    assert_eq!(info, info2);
 }
 
 fn create_unique_file_name(function_name: &str) -> String {
