+++ duva/tests/cluster_ops/mod.rs
@@ -1,0 +1,6 @@
+mod test_cluster_forget_makes_all_nodes_forget_target_node;
+mod test_cluster_forget_when_wrong_id_given;
+mod test_cluster_known_nodes_increase_when_new_replica_is_added;
+mod test_disseminate_peers;
+mod test_heartbeat;
+mod test_removes_node_when_heartbeat_is_not_received_for_certain_time;
