+++ duva/tests/cluster_ops/test_cluster_forget_makes_all_nodes_forget_target_node.rs
@@ -1,5 +1,5 @@
-mod common;
-use common::{ServerEnv, array, check_internodes_communication, spawn_server_process};
+use crate::common::{ServerEnv, array, check_internodes_communication, spawn_server_process};
+
 use duva::{clients::ClientStreamHandler, domains::query_parsers::query_io::QueryIO};
 
 #[tokio::test]
