+++ duva/tests/common.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code, unused_variables)]
+
 use bytes::Bytes;
 use duva::domains::query_parsers::query_io::QueryIO;
 use duva::make_smart_pointer;
